services/shop/requirements.txt
python-dotenv==1.0.0
fastapi==0.109.*
uvicorn[standard]
python-multipart==0.0.*

mysqlclient
sqlalchemy